Mudança na variável n de int para long

+++ trabalho-2/eratostenes.c
@@ -3,15 +3,15 @@
 #include <stdbool.h>
 #include <omp.h>
 
-int eratostenes(int n);
+long eratostenes(long n);
 
 
 int main(int argc, char* argv[]) {
-    int n = -1;
-    int qnt_primos = -1;
+    long n = -1;
+    long qnt_primos = -1;
 
     if (argc == 2) {
-        n = atoi(argv[1]);
+        n = atol(argv[1]);
     } else {
         fprintf(stderr, "Número de argumentos inválido!\n");
         exit(1);
@@ -23,14 +23,14 @@ int main(int argc, char* argv[]) {
 
     double t_fim = omp_get_wtime();
 
-    printf("Existem %d números primos entre 2 e %d\n", qnt_primos, n);
+    printf("Existem %ld números primos entre 2 e %ld\n", qnt_primos, n);
     printf("Tempo de execução: %.15lf\n", t_fim - t_inicio);
 
     return 0;
 }
 
 
-int eratostenes(int n) {
+long eratostenes(long n) {
     if (n < 2)
         return 0;
 
@@ -42,23 +42,23 @@ int eratostenes(int n) {
     }
 
     #pragma omp parallel for default(none) shared(n, eh_primo) schedule(static)
-    for (int i = 0; i < n + 1; ++i) {
+    for (long i = 0; i < n + 1; ++i) {
         eh_primo[i] = true;
     } // barreira implícita
 
-    for (int i = 2; i*i <= n; ++i) {
+    for (long i = 2; i*i <= n; ++i) {
         if (eh_primo[i]) {
             #pragma omp parallel for default(none) shared(n, i, eh_primo) schedule(static)
-            for (int j = i*i; j <= n; j += i) {
+            for (long j = i*i; j <= n; j += i) {
                 eh_primo[j] = false;
             } // barreira implícita            
         }            
     }        
 
-    int qnt_primos = 0;
+    long qnt_primos = 0;
 
     #pragma omp parallel for default(none) shared(n, eh_primo) reduction(+:qnt_primos) schedule(static)
-    for (int i = 3; i <= n; i += 2) {
+    for (long i = 3; i <= n; i += 2) {
         if (eh_primo[i])
             ++qnt_primos;
     } // barreira implícita
